PythonProject18/Crypto News/quart_app/requirements.txt
quart
motor
python-dotenv
